Time: 4 ms (33.33%), Space: 6.2 MB (66.67%) - LeetHub

+++ 2243-calculate-digit-sum-of-a-string/2243-calculate-digit-sum-of-a-string.cpp
@@ -2,11 +2,11 @@ class Solution {
 
 public:
     string digitSum(string s, int k) {
-        while ((int) s.size() > k) {
+        while ( s.size() > k) {
             int sum = 0;
             string t = "";
             int count = 0;
-            for (int i = 0; i < (int) s.size(); i++) {
+            for (int i = 0; i <  s.size(); i++) {
                 sum += (s[i] - '0');
                 ++count;
                 if (count == k) {
